new sprites added / new move functions

+++ Game.js
@@ -90,7 +90,7 @@
 
 		player = objectPool.getPlayer();
 		player.startMe();
-		player.setWeapon("double");
+		player.setWeapon("single");
 
 
 		// game event listener for all events that control gameplay
